+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # stoker
 
+## 0.4.58
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.20
+    - @stoker-platform/cli@0.5.43
+    - @stoker-platform/web-app@0.5.49
+    - @stoker-platform/node-client@0.5.32
+    - @stoker-platform/utils@0.5.26
+    - @stoker-platform/web-client@0.5.28
+
 ## 0.4.57
 
 ### Patch Changes