+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/node-client
 
+## 0.5.32
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.20
+    - @stoker-platform/utils@0.5.26
+
 ## 0.5.31
 
 ### Patch Changes