+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stoker-platform/utils
 
+## 0.5.26
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.20
+
 ## 0.5.25
 
 ### Patch Changes