+++ packages/web-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/web-client
 
+## 0.5.28
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.20
+    - @stoker-platform/utils@0.5.26
+
 ## 0.5.27
 
 ### Patch Changes